pyproject.toml
[project]
name = "stock-fast-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.10.10",
    "beautifulsoup4>=4.12.3",
    "fake-useragent>=1.5.1",
    "fastapi>=0.115.0",
    "lxml>=5.3.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "seria-library[tortoise]>=1.5.4",
    "tortoise-orm[asyncpg]>=0.21.6",
    "uvicorn>=0.31.0",
]
license = { file = "LICENSE" }
authors = [{ name = "seriaati", email = "seria.ati@gmail.com" }]

[tool.pyright]
enableTypeIgnoreComments = false
reportIncompatibleMethodOverride = false
reportIncompatibleVariableOverride = false
reportUnnecessaryComparison = true
reportUnnecessaryContains = true
reportUnnecessaryIsInstance = true
reportUnnecessaryTypeIgnoreComment = true
typeCheckingMode = "standard"
